pyproject.toml
[tool.poetry]
name = "instagramproject"
version = "0.1.0"
description = ""
authors = ["Khaliq Salawou <khaliqsalawoudeen@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
httpx = "0.26"
pandas = "^2.2.3"
selectolax = "^0.3.27"
googlesearch-python = "^1.2.5"
ipykernel = "^6.29.5"
requests = "^2.32.3"
pillow = "^11.0.0"
pytesseract = "^0.3.13"
playwright = "^1.49.0"
opencv-python = "^4.10.0.84"
supabase = "^2.10.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
